requirements.txt
streamlit~=1.31.1
google-generativeai
python-dotenv~=1.0.1